Feature_based/requirements.txt
streamlit>=1.28.0
torch>=2.0.0
kornia>=0.8.0
kornia-moons>=0.2.9
opencv-python>=4.8.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=10.0.0

